Back to a default template
This is just for making easier to know that this is Unlicensed.
The corrections I made earlier were just for making clear this is code and not software, but GitHub can not recognize my license as "Unlicensed" that way. Therefore, I confirm this is Unlicensed, 100% free code (and/or software).